+++ CHANGELOG.md
@@ -13,3 +13,8 @@
 ## [2.1.0] - 2024-09-22
 ### New Feature
 - Email Optimise.
+
+## [2.2.0] - 2024-09-22
+### New Feature
+- Merge and Minify multiple assets.
+- Minify asset.